pyproject.toml
[project]
name = "tecno-etl"
version = "0.2.0"
description = "Pipeline ETL moderno para datos de TecnoMundo con validación robusta (AWS Serverless)"
authors = [
    { name = "Franco Aguilera", email = "franco18min@github.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.4",
    "numpy>=1.26.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.2",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

# Scripts eliminados (CLI ya no existe)
# [project.scripts]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tecno_etl"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Removido --cov para evitar errores si no está instalado
addopts = "-v"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
